requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
sqlparse==0.4.2
python-dotenv==0.15.0
psycopg2==2.8.5
djangorestframework==3.11.1
pytz==2020.1

